requirements.txt
pandas>=2.0.0
datasets>=2.14.0
openai>=1.0.0
requests>=2.31.0
pyarrow>=13.0.0
kaggle>=1.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
ratelimit>=2.2.1
